+++ examples/Buttons/Button.md
@@ -1,12 +1,12 @@
 ```jsx
 import { VARIANT } from 'Theme/MarkOneTheme';
 
-<BasicButton
+<Button
   clickHandler={function() {
     alert('You clicked the basic button')
   }}
   variant={VARIANT.BASE}
 >
   Click Me!
-</BasicButton>
+</Button>
 ```
