+++ src/Buttons/index.ts
@@ -1,2 +1,2 @@
-export { default as BasicButton } from './BasicButton';
+export { default as Button } from './Button';
 export { default as BorderlessButton } from './BorderlessButton';
