+++ src/Theme/MarkOneTheme.ts
@@ -69,21 +69,25 @@ const MarkOneTheme: BaseTheme = {
         light: WHITE,
         medium: darken(0.175, WHITE),
         dark: '#93a1ad',
+        text: BLACK,
       },
       info: {
         light: '#95b5df',
         medium: '#4e88c7',
         dark: '#084161',
+        text: WHITE,
       },
       positive: {
         light: '#81e981',
         medium: '#4db848',
         dark: '#178717',
+        text: WHITE,
       },
       negative: {
         light: '#ff8f9c',
         medium: '#ff4040',
         dark: '#ed1b2a',
+        text: WHITE,
       },
     },
     text: {
